# Conflicts:
#	src/main/resources/templates/propagandaBusiness/propagandaInformation/query.html

+++ src/main/java/com/anonymous/service/PropagandaMaterialsRecipientsService.java
@@ -1,7 +1,6 @@
 package com.anonymous.service;
 
 import com.anonymous.domain.PropagandaMaterials;
-import com.anonymous.domain.PropagandaMaterialsProduced.ApprovalStatus;
 import com.anonymous.domain.PropagandaMaterialsRecipients;
 import com.anonymous.domain.User;
 import com.anonymous.repository.PropagandaMaterialsRecipientsRepository;
@@ -154,13 +153,24 @@ public class PropagandaMaterialsRecipientsService implements PropagandaMaterials
             }
         }
         for (String id : ids) {
-            PropagandaMaterialsRecipients propagandaMaterialsRecipients = getPropagandaMaterialsRecipients(id);
+            if (!deleted(id)) {
+                throw new RuntimeException("删除 宣传物资领用 出现异常 ");
+            }
+        }
+        return true;
+    }
+
+    private boolean deleted(String id) {
+        PropagandaMaterialsRecipients propagandaMaterialsRecipients = getPropagandaMaterialsRecipients(id);
+        if (propagandaMaterialsRecipients != null) {
             for (PropagandaMaterials propagandaMaterials : propagandaMaterialsRecipients.getPropagandaMaterials()) {
-                propagandaMaterialsService.delete(propagandaMaterials);
+                if (!propagandaMaterialsService.delete(propagandaMaterials)) {
+                    throw new RuntimeException("删除 宣传物资领用 链接删除 宣传物资内容 出现异常 ");
+                }
             }
             propagandaMaterialsRecipientsRepository.delete(id);
         }
-        return true;
+        return propagandaMaterialsRecipientsRepository.findOne(id) == null ? true : false;
     }
 
     //宣传物资领用统计报表
